warning cleanup
